move pymor.functions to pymor.analyticalproblems.functions
and flatten to single module
