+++ src/pymor/analyticalproblems/burgers.py
@@ -7,7 +7,7 @@ import numpy as np
 from pymor.analyticalproblems.elliptic import StationaryProblem
 from pymor.analyticalproblems.domaindescriptions import LineDomain, RectDomain, TorusDomain, CircleDomain
 from pymor.analyticalproblems.instationary import InstationaryProblem
-from pymor.functions.basic import ConstantFunction, ExpressionFunction
+from pymor.analyticalproblems.functions import ConstantFunction, ExpressionFunction
 from pymor.parameters.spaces import CubicParameterSpace
