+++ src/pymor/analyticalproblems/helmholtz.py
@@ -4,7 +4,7 @@
 
 from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
-from pymor.functions.basic import ConstantFunction, LincombFunction
+from pymor.analyticalproblems.functions import ConstantFunction, LincombFunction
 from pymor.parameters.functionals import ExpressionParameterFunctional
 from pymor.parameters.spaces import CubicParameterSpace
