+++ src/pymor/analyticalproblems/thermalblock.py
@@ -6,7 +6,7 @@ from itertools import product
 
 from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
-from pymor.functions.basic import ConstantFunction, ExpressionFunction, LincombFunction
+from pymor.analyticalproblems.functions import ConstantFunction, ExpressionFunction, LincombFunction
 from pymor.parameters.functionals import ProjectionParameterFunctional
 from pymor.parameters.spaces import CubicParameterSpace
