+++ src/pymor/basic.py
@@ -25,6 +25,8 @@ from pymor.analyticalproblems.burgers import burgers_problem, burgers_problem_2d
 from pymor.analyticalproblems.domaindescriptions import RectDomain, CylindricalDomain, TorusDomain, LineDomain, CircleDomain
 from pymor.analyticalproblems.domaindescriptions import DiscDomain, CircularSectorDomain, PolygonalDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.analyticalproblems.functions import (ConstantFunction, GenericFunction, ExpressionFunction, LincombFunction,
+                                                BitmapFunction)
 from pymor.analyticalproblems.helmholtz import helmholtz_problem
 from pymor.analyticalproblems.instationary import InstationaryProblem
 from pymor.analyticalproblems.thermalblock import thermal_block_problem
@@ -43,10 +45,6 @@ from pymor.discretizers.builtin import (discretize_stationary_cg, discretize_ins
                                         OnedGrid, TriaGrid, RectGrid, load_gmsh)
 from pymor.discretizers.builtin.domaindiscretizers.default import discretize_domain_default
 from pymor.discretizers.builtin.grids.boundaryinfos import EmptyBoundaryInfo, GenericBoundaryInfo, AllDirichletBoundaryInfo
-
-from pymor.functions.basic import ConstantFunction, GenericFunction, ExpressionFunction, LincombFunction
-from pymor.functions.bitmap import BitmapFunction
-
 
 from pymor.operators.constructions import (LincombOperator, Concatenation, ComponentProjection, IdentityOperator,
                                            ConstantOperator, ZeroOperator, VectorArrayOperator, VectorOperator,
