+++ src/pymordemos/elliptic.py
@@ -28,8 +28,8 @@ import numpy as np
 
 from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.analyticalproblems.functions import ExpressionFunction, ConstantFunction
 from pymor.discretizers.builtin import discretize_stationary_cg, discretize_stationary_fv, RectGrid, TriaGrid
-from pymor.functions.basic import ExpressionFunction, ConstantFunction
 
 
 def elliptic_demo(args):
