+++ src/pymordemos/elliptic2.py
@@ -21,8 +21,8 @@ from docopt import docopt
 
 from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.analyticalproblems.functions import ExpressionFunction, LincombFunction, ConstantFunction
 from pymor.discretizers.builtin import discretize_stationary_cg, discretize_stationary_fv
-from pymor.functions.basic import ExpressionFunction, LincombFunction, ConstantFunction
 from pymor.parameters.functionals import ProjectionParameterFunctional, ExpressionParameterFunctional
 from pymor.parameters.spaces import CubicParameterSpace
