+++ src/pymordemos/elliptic_oned.py
@@ -21,8 +21,8 @@ from docopt import docopt
 
 from pymor.analyticalproblems.domaindescriptions import LineDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.analyticalproblems.functions import ExpressionFunction, ConstantFunction, LincombFunction
 from pymor.discretizers.builtin import discretize_stationary_cg, discretize_stationary_fv
-from pymor.functions.basic import ExpressionFunction, ConstantFunction, LincombFunction
 from pymor.parameters.functionals import ProjectionParameterFunctional, ExpressionParameterFunctional
 from pymor.parameters.spaces import CubicParameterSpace
