+++ src/pymortests/fixtures/analyticalproblem.py
@@ -7,9 +7,9 @@ import pytest
 from pymor.analyticalproblems.burgers import burgers_problem, burgers_problem_2d
 from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.analyticalproblems.functions import GenericFunction, ConstantFunction, LincombFunction
 from pymor.analyticalproblems.helmholtz import helmholtz_problem
 from pymor.analyticalproblems.thermalblock import thermal_block_problem
-from pymor.functions.basic import GenericFunction, ConstantFunction, LincombFunction
 from pymor.parameters.functionals import ExpressionParameterFunctional
